Backend/requirements.txt
# FastAPI and related packages
fastapi==0.116.1
uvicorn==0.32.1

# Database
sqlalchemy==2.0.43
pymysql==1.1.1
mysqlclient==2.2.7

# Data validation
pydantic==2.11.7

# Authentication
pyjwt==2.10.1

# Email
fastapi-mail==1.4.1

# HTTP client
httpx==0.28.1

# Environment variables
python-dotenv==1.0.1

